requirements.txt
fastapi==0.121.1
uvicorn[standard]==0.38.0
paho-mqtt==2.1.0
python-socketio==5.14.3
jinja2==3.1.6
aiofiles==24.1.0
python-multipart==0.0.20
tensorflow==2.20.0
pillow==12.0.0
